refactor(adapters): extract helpers to fix max-lines lint
